Add retrieval via Selenium in Lambda Layer
**NOTE: Abandoned in place** This code worksßat least in a technical
sense. It does retrieve content from the Congressional Research Service
website using a headless Chromium driven by Selenium in an AWS Lambda
layer. _But_, come to find out, the CRS is fronted by Cloudflare, which
is designed to prevent exactly what it is I'm trying to do.

I'm abandoning this part of the code in place because I'll probably
want to refer to it for a later project. Ongoing efforts will just run
locally.